python/orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pathflow-guard-orchestrator"
version = "0.2.0"
description = "Decision engine for digital pathology slide-ingestion QC."
license = { text = "Apache-2.0" }
requires-python = ">=3.12"
keywords = ["digital-pathology", "quality-control", "medical-software", "openslide"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "openslide-python>=1.4.3,<2",
    "openslide-bin>=4,<5; platform_system == 'Windows'",
    "pillow>=12,<13",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6,<7",
    "ruff>=0.11,<0.12",
]

[project.scripts]
pathflow-guard = "pathflow_guard.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
